requirements.txt
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-html>=3.2.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-rerunfailures>=12.0
locust>=2.15.0
python-dotenv>=1.0.0
allure-pytest>=2.13.0
pymysql>=1.1.0

# Code quality tools
flake8>=6.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
radon>=6.0.0
flake8-docstrings>=1.7.0
flake8-bugbear>=23.0.0
flake8-comprehensions>=3.14.0
mccabe>=0.7.0
